// src/gen/api/types/ListPostsByUser.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Post } from './Post.ts'

export type ListPostsByUserPathParams = {
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  id: string
}

export type ListPostsByUserQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated posts by author
 */
export type ListPostsByUser200 = {
  /**
   * @type array
   */
  data: Post[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListPostsByUser400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListPostsByUser401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListPostsByUser500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListPostsByUserQueryResponse = ListPostsByUser200

export type ListPostsByUserQuery = {
  Response: ListPostsByUser200
  PathParams: ListPostsByUserPathParams
  QueryParams: ListPostsByUserQueryParams
  Errors: ListPostsByUser400 | ListPostsByUser401 | ListPostsByUser500
}